pyproject.toml
[tool.poetry]
name = "analyze-agent"
version = "0.1.0"
description = ""
authors = ["crazyivan1289 <crazyivan1289@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.125.0"
slowapi = "^0.1.9"
langchain = "^0.3.27"
langgraph = "^0.6.7"
langchain-openai = "^0.3.19"
langchain-community = "^0.3.29"
langchain-huggingface = "^0.1.2"
gigachat = "^0.1.43"
fastmcp = "^2.14.1"
tarantool = "^1.2.0"
httpx = {extras = ["http2"], version = "^0.28.1"}
tenacity = "^9.1.2"
python-dotenv = "^1.1.1"
hvac = "^2.1.0"
pyyaml = "^6.0.1"
xmltodict = "^0.15.0"
aiofiles = "^24.1.0"
streamlit = "1.45.0"
pydantic-settings = "^2.12.0"
uvicorn = "^0.38.0"
fpdf2 = "^2.8.1"
rich = "^14.0.0"
apscheduler = "^3.10.4"
msgpack = "^1.1.0"
opentelemetry-api = "^1.39.1"
opentelemetry-sdk = "^1.39.1"
opentelemetry-instrumentation-fastapi = "^0.60b1"
tavily-python = "^0.7.17"
openai = "^2.13.0"
faststream = {extras = ["rabbit"], version = "^0.6.4"}
starlette = "^0.50.0"
jinja2 = "^3.1.6"
cryptography = "^46.0.3"
setuptools = "^80.9.0"
watchdog = "^6.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.11"
pyright = "^1.1.404"
detect-secrets = "^1.5.0"
pylint = "^3.3.8"
pip-audit = "^2.10.0"
bandit = "^1.8.6"
deptry = "^0.23.1"
black = "^25.1.0"
vulture = "^2.14"
pytest = "^8.4.1"
pytest-asyncio = "^1.3.0"
mypy = "^1.12.0"
types-requests = "^2.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.vulture]
ignore_names = [
    "x_api_key"
]
min_confidence = 70
exclude = [
    "*/migrations/*",
    "*/tests/*",
    "*/venv/*"
]

[tool.pyright]
include = ["app"]
exclude = ["**/__pycache__"]
typeCheckingMode = "basic"
pythonVersion = "3.12"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
  "F",   # pyflakes
  "E",   # pycodestyle
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
]
ignore = ["E501"]
per-file-ignores = { "__init__.py" = ["F401"] }

[tool.ruff.format]
line-ending = "auto"
indent-style = "space"
quote-style = "double"
